Storing Citations and Citations Metadata
Based on the schema discussed in ticket 18, the current metrics-service contains functionality for storing the cityations metadata which could be used in the citations counts as well as the citations list.

The updateCitationMetadata() function should be called via CRON to support the functionality of updating the metadata om a regular basis.

Reference #18 #19