Creating a new session only once per batch
In the pid_resolution module we cerate a new Session for every pid that we have in the batch that needs some processing.

Not really sure if I'm heading in the right direction - but would it be more efficient to create a single session for an entire batch? (if it does not have any significant impact on performance then we should leave as it is and revert back this commit)

@datadavev could you please review this commit?